Set source association when setting a has_one through record
When a through association that is a has_one association is set, we
don't set the source association, which might be confusing, because the
source association is available after saving and reloading the record.

For example:

  class User < ActiveRecord::Base
    has_one :subscription
    has_one :plan, through: :subscription
  end

  class Plan < ActiveRecord::Base
    has_many :subscriptions
  end

  class Subscription < ActiveRecord::Base
    belongs_to :plan
  end

The subscription association in this case is a has_one association and
also a through association. Before this commit the result of assigning
the subscription to the user would be the following:

  user = User.new
  plan = Plan.new(name: "Startup plan")
  subscription = Subscription.new(plan: plan)

  user.subscription = subscription

  user.plan #=> nil

  user.save
  user.reload

  user.plan #=> #<Plan name: "Startup plan">

With a fix introduced in the commit the plan would be set before saving
the record:

  user = User.new
  plan = Plan.new(name: "Startup plan")
  subscription = Subscription.new(plan: plan)

  user.subscription = subscription

  user.plan #=> #<Plan name: "Startup plan">

  user.save
  user.reload

  user.plan #=> #<Plan name: "Startup plan">